move src/ast-cfg-commands to language-support/ast-viewer/ast-cfg-commands
